requirements.txt
# Core web framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.28.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# AI and ML
google-generativeai>=0.3.0
pinecone>=4.0.0
numpy>=1.21.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0

# Development and testing (optional)
pytest>=7.0.0
httpx>=0.24.0
black>=23.0.0
isort>=5.12.0
